+++ lihil/plugins/testclient.py
@@ -352,7 +352,7 @@ class LocalClient:
         headers: dict[str, str] | None = None,
     ) -> RequestResult:
         """
-        # TODO: override ep dependencies
+        TODO: override ep dependencies
         1. make a new graph, merge ep.graph
         2. override in the new graph
         3. set ep.graph = new graph